Move storybook react config to react subdirectory
